+++ backend/internal/usecase/menu_usecase/create_item.go
@@ -20,7 +20,7 @@ type CreateMenuItemInput struct {
 	Description   string                   `json:"description"`
 	Image         string                   `json:"image"`
 	Available     bool                     `json:"available"`
-	BaseUnitPrice int                      `json:"baseUnitPrice"`
+	BaseUnitPrice int64                    `json:"baseUnitPrice"`
 	Options       []*entity.MenuItemOption `json:"options"`
 	Categories    []int                    `json:"categories"`
 }
